Testing
Moved useful fixtures to conftest file for reusability.
Minor refactoring for better readability and consistency.
